Factored out setup into traits
