+++ twitter_classifier/scala/src/main/scala/com/databricks/apps/twitterClassifier/Collect.scala
@@ -1,16 +1,14 @@
 package com.databricks.apps.twitterClassifier
 
 import com.google.gson.Gson
-import org.apache.spark.SparkContext
-import org.apache.spark.sql.SparkSession
 import org.apache.spark.streaming.dstream.DStream
 import org.apache.spark.streaming.twitter.TwitterUtils
-import org.apache.spark.streaming.{Seconds, StreamingContext}
 
 /** Collect at least the specified number of tweets into json text files. */
 object Collect extends App {
   val options = CollectOptions.parse(args)
-  import options._
+  import com.databricks.apps.twitterClassifier.Collect.options._
+  val intervalInSecs = options.intervalInSecs
 
   if (tweetDirectory.exists) {
     if (options.overWrite) { // nuke any existing files
@@ -21,19 +19,10 @@ object Collect extends App {
     }
   } else
     tweetDirectory.mkdirs()
+  new Collect(options)
+}
 
-  println("Initializing Streaming Spark Context...")
-  val spark = SparkSession
-    .builder
-    .appName(getClass.getSimpleName.replace("$", ""))
-    .master("local[*]")
-    .getOrCreate()
-
-  val sc: SparkContext = spark.sparkContext
-  // Suppress "WARN BlockManager: Block input-0-1478266015800 replicated to only 0 peer(s) instead of 1 peers" messages
-  sc.setLogLevel("ERROR")
-  val ssc = new StreamingContext(sc, Seconds(intervalInSecs))
-
+class Collect(options: CollectOptions) extends StreamingSessionLike {
   val tweetStream: DStream[String] = TwitterUtils.createStream(ssc, Utils.maybeTwitterAuth)
     .map(new Gson().toJson(_))
