+++ twitter_classifier/scala/src/main/scala/com/databricks/apps/twitterClassifier/Predict.scala
@@ -1,37 +1,16 @@
 package com.databricks.apps.twitterClassifier
 
-import org.apache.spark.SparkContext
 import org.apache.spark.mllib.clustering.KMeansModel
 import org.apache.spark.mllib.linalg.Vector
-import org.apache.spark.sql.SparkSession
 import org.apache.spark.streaming.twitter._
-import org.apache.spark.streaming.{Seconds, StreamingContext}
 
 object Predict extends App {
   val options = PredictOptions.parse(args)
   new Predict(options).predict()
 }
 
-trait SparkSessionLike {
-  def intervalInSecs: Int
-
-  val spark = SparkSession
-    .builder
-    .appName(getClass.getSimpleName.replace("$", ""))
-    .master("local[*]")
-    .getOrCreate()
-
-  val sc: SparkContext = spark.sparkContext
-  // Suppress "WARN BlockManager: Block input-0-1478266015800 replicated to only 0 peer(s) instead of 1 peers" messages
-  sc.setLogLevel("ERROR")
-  val ssc = new StreamingContext(sc, Seconds(intervalInSecs))
-
-  // Suppress "WARN BlockManager: Block input-0-1478266015800 replicated to only 0 peer(s) instead of 1 peers" messages
-  ssc.sparkContext.setLogLevel("ERROR")
-}
-
 /** Pull live tweets and filter them for tweets in the chosen cluster */
-class Predict(options: PredictOptions) extends SparkSessionLike {
+class Predict(options: PredictOptions) extends StreamingSessionLike {
   import options._
   val intervalInSecs = options.intervalInSecs
